add data base
include EEPROM
add structur Color
add getData

+++ Lab 4/Main/include/color.h
@@ -4,7 +4,7 @@
 #include <Arduino.h>
 
 struct Color {
-    uint8_t initFlag = 0;
+    uint16_t initFlag = 0;
     uint8_t red;
     uint8_t green;
     uint8_t blue;
